refine code

+++ x.sh
@@ -3,6 +3,15 @@
 #set -o trace
 
 #------------------------------------------------------------------------------
+#	Information
+#------------------------------------------------------------------------------
+
+author="James Chan"
+email="ccchan0109@gmail.com"
+version="1.1.1"
+updateDate="2022/11/22"
+
+#------------------------------------------------------------------------------
 #	Parameters
 #------------------------------------------------------------------------------
 
@@ -11,6 +20,13 @@ UNLINK=no
 BOOTSTRAP=no
 ALIAS=no
 GITPROMPT=no
+
+CONFIGS=(
+	".vimrc"
+	".tmux.conf"
+	".gitconfig"
+	".gitignore"
+)
 
 #------------------------------------------------------------------------------
 #	Functions
@@ -30,29 +46,43 @@ OPTIONS:
 	-h		Show this usage
 
 AUTHOR:
-	James Chan - ccchan0109@gmail.com
+	$author - $email
 
 VERSION:
-	1.1.0 @ 2022/11/4
-	1.0.2 @ 2019/10/8
-	1.0.1 @ 2017/8/24
+	$version @ $updateDate
 
 EOF
 }
 
 bootstrap()
 {
-	sudo apt update
-	sudo apt install -y ssh vim curl ctags cscope make tmux sed silversearcher-ag cifs-utils
+	install_dependent_packages
 
-	curl https://raw.githubusercontent.com/Shougo/neobundle.vim/master/bin/install.sh | sh
+	install_vim_neobundle
 
-	link
+	link_configs
 
+	echo "don't forget to set up git user name and email"
 	echo "please relogin again to apply the setting"
 }
 
-install_prompt()
+install_dependent_packages()
+{
+	if [ "$EUID" == "0" ]; then
+		apt update
+		apt install -y ssh vim curl exuberant-ctags cscope make tmux sed silversearcher-ag cifs-utils
+	else
+		sudo apt update
+		sudo apt install -y ssh vim curl exuberant-ctags cscope make tmux sed silversearcher-ag cifs-utils
+	fi
+}
+
+install_vim_neobundle()
+{
+	curl https://raw.githubusercontent.com/Shougo/neobundle.vim/master/bin/install.sh | sh
+}
+
+install_git_prompt()
 {
 	if [ ! -d ~/.bash-git-prompt ]; then
 		git clone https://github.com/magicmonty/bash-git-prompt.git ~/.bash-git-prompta
@@ -62,7 +92,7 @@ install_prompt()
 
 add_alias()
 {
-	if [ ! -d ~/.bash_aliases ]; then
+	if [ ! -f ~/.bash_aliases ]; then
 		echo "alias tmux='TERM=screen-256color-bce tmux'" >> ~/.bash_aliases
 		echo "alias tt='tmux new -s'" >> ~/.bash_aliases
 		echo "alias ttk='tmux kill-session -t'" >> ~/.bash_aliases
@@ -75,28 +105,56 @@ add_alias()
 
 		# update to .bashrc
 		echo "source ~/.bash_aliases" >> ~/.bashrc
+	else
+		echo ".bash_aliases is already existed"
 	fi
 }
 
 link() {
-	echo "Linking config files in env/"
+	config=$1
 
-	configs=$(find env/ -maxdepth 1 -xtype f)
-	for config in $configs; do
-		config=$(basename $config)
+	if [ -L ~/$config ] && [ -e ~/$config ]; then
+		echo "$config is already linked"
+	else
 		[ -f ~/$config ] && [ ! -f ~/${config}.bak ] && mv -f ~/$config ~/${config}.bak && echo "backup config $config to ${config}.bak"
 		ln -sf $(pwd)/env/$config ~/$config
+		echo "link of $config is built"
+	fi
+}
+
+link_configs() {
+	echo "linking config files..."
+
+	for config in ${CONFIGS[@]}; do
+		link $config
 	done
+
+	echo "linking config files done"
 }
 
 unlink() {
-	echo "Restore config files"
+	config=$1
 
-	configs=$(find env/ -maxdepth 1 -xtype f)
-	for config in $configs; do
-		config=$(basename $config)
-		rm -f ~/$config && mv -f ~/${config}.bak ~/$config
+	if [ -L ~/$config ] && [ -e ~/$config ]; then
+		rm -f ~/$config
+		if [ -f ~/${config}.bak ]; then
+			mv -f ~/${config}.bak ~/$config && echo "restore config $config from ~/${config}.bak"
+		else
+			echo "link of $config is removed"
+		fi
+	else
+		echo "$config link is not existed"
+	fi
+}
+
+unlink_configs() {
+	echo "restore config files..."
+
+	for config in ${CONFIGS[@]}; do
+		unlink $config
 	done
+
+	echo "restore config files done"
 }
 
 #------------------------------------------------------------------------------
@@ -136,13 +194,13 @@ Main() {
 	if [ "$BOOTSTRAP" == "yes" ]; then
 		bootstrap
 	elif [ "$LINK" == "yes" ]; then
-		link
+		link_configs
 	elif [ "$UNLINK" == "yes" ]; then
-		unlink
+		unlink_configs
 	elif [ "$ALIAS" == "yes" ]; then
 		add_alias
 	elif [ "$GITPROMPT" == "yes" ]; then
-		install_prompt
+		install_git_prompt
 	else
 		usage
 	fi
